esame-16-giugno-2023/src/requirements.txt
flask
flask-restful
WTForms
requests
google-cloud-firestore
gunicorn
python-dateutil